LayoutTests/css3/scroll-snap/scroll-snap-iframe-expected.txt

iframe processed all tests successfully.
PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
Verify that the iframe propery parses snap-point settings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
